more misc cleanup and removing some usings that should have been removed when moving things from one project to another

+++ BizHawk.Emulation.Common/CPUs/CP1610/Execute.cs
@@ -185,7 +185,7 @@ namespace BizHawk.Emulation.Common.Components.CP1610
 							// Unknown opcode.
 							throw new ArgumentException();
 					}
-					RegisterPC = (ushort)addr;
+					RegisterPC = addr;
 					cycles = 12;
 					Interruptible = true;
 					break;
